Moving persist_rollback_info to deploy.py
This may sound counter-intuitive but makes sense. The deploy script is
responsible for writing rollback information whereas rollback.py reads
rollback information in order to generate the rollback script.

deploy.py provides the input for rollback.py.
